refactor(memory): modularize core memory layer
Split the large memory implementation into focused utility, database, scoring, and operation modules while keeping the MARMMemory facade intact.

Update affected tests for the new module boundaries, start the v2.14.0 changelog entry, refresh contributor architecture docs, and remove low-value announcement workflows pending a better automation path.
