--- .github/workflows/discord-announcements.yml
@@ -1,77 +1,0 @@
-name: Discord Announcements
-
-on:
-  push:
-    tags:
-      - "v*"
-
-permissions: {}
-
-jobs:
-  announce-release:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Check webhook configuration
-        id: webhook
-        env:
-          DISCORD_WEBHOOK_URL: ${{ secrets.DISCORD_ANNOUNCEMENTS_WEBHOOK_URL }}
-        run: |
-          if [ -z "$DISCORD_WEBHOOK_URL" ]; then
-            echo "enabled=false" >> "$GITHUB_OUTPUT"
-          else
-            echo "enabled=true" >> "$GITHUB_OUTPUT"
-          fi
-
-      - name: Post release announcement to Discord
-        if: ${{ steps.webhook.outputs.enabled == 'true' }}
-        env:
-          DISCORD_WEBHOOK_URL: ${{ secrets.DISCORD_ANNOUNCEMENTS_WEBHOOK_URL }}
-          RELEASE_TAG: ${{ github.ref_name }}
-          RELEASE_URL: ${{ github.server_url }}/${{ github.repository }}/releases/tag/${{ github.ref_name }}
-          REPO: ${{ github.repository }}
-        run: |
-          python - <<'PY'
-          import json
-          import os
-          import sys
-          import urllib.request
-          import urllib.error
-
-          payload = {
-              "username": "MARM Releases",
-              "embeds": [
-                  {
-                      "title": f"{os.environ['RELEASE_TAG']} tagged",
-                      "url": os.environ["RELEASE_URL"],
-                      "description": "New version tag pushed for MARM.",
-                      "color": 5814783,
-                      "fields": [
-                          {
-                              "name": "Repository",
-                              "value": os.environ["REPO"],
-                              "inline": True,
-                          },
-                          {
-                              "name": "Tag",
-                              "value": os.environ["RELEASE_TAG"],
-                              "inline": True,
-                          },
-                      ],
-                  }
-              ],
-          }
-
-          req = urllib.request.Request(
-              os.environ["DISCORD_WEBHOOK_URL"],
-              data=json.dumps(payload).encode("utf-8"),
-              headers={"Content-Type": "application/json"},
-          )
-          try:
-              urllib.request.urlopen(req).read()
-          except urllib.error.HTTPError as exc:
-              print(f"Discord webhook post failed: HTTP {exc.code} {exc.reason}", file=sys.stderr)
-              sys.exit(0)
-          except urllib.error.URLError as exc:
-              print(f"Discord webhook post failed: {exc}", file=sys.stderr)
-              sys.exit(0)
-          PY
