--- .github/workflows/discord-get-involved.yml
@@ -1,156 +1,0 @@
-name: Discord Get Involved
-
-on:
-  pull_request:
-    types: [opened, closed]
-  issues:
-    types: [opened, labeled]
-
-permissions: {}
-
-jobs:
-  post-get-involved:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Check webhook configuration
-        id: webhook
-        env:
-          DISCORD_WEBHOOK_URL: ${{ secrets.DISCORD_GET_INVOLVED_WEBHOOK_URL }}
-        run: |
-          if [ -z "$DISCORD_WEBHOOK_URL" ]; then
-            echo "enabled=false" >> "$GITHUB_OUTPUT"
-          else
-            echo "enabled=true" >> "$GITHUB_OUTPUT"
-          fi
-
-      - name: Post contributor activity to Discord
-        if: ${{ steps.webhook.outputs.enabled == 'true' }}
-        env:
-          DISCORD_WEBHOOK_URL: ${{ secrets.DISCORD_GET_INVOLVED_WEBHOOK_URL }}
-          EVENT_NAME: ${{ github.event_name }}
-          ACTION_NAME: ${{ github.event.action }}
-          REPO: ${{ github.repository }}
-          ACTOR: ${{ github.actor }}
-          PR_AUTHOR: ${{ github.event.pull_request.user.login }}
-          PR_TITLE: ${{ github.event.pull_request.title }}
-          PR_URL: ${{ github.event.pull_request.html_url }}
-          PR_NUMBER: ${{ github.event.pull_request.number }}
-          PR_MERGED: ${{ github.event.pull_request.merged }}
-          ISSUE_AUTHOR: ${{ github.event.issue.user.login }}
-          ISSUE_TITLE: ${{ github.event.issue.title }}
-          ISSUE_URL: ${{ github.event.issue.html_url }}
-          ISSUE_NUMBER: ${{ github.event.issue.number }}
-          LABEL_NAME: ${{ github.event.label.name }}
-        run: |
-          python - <<'PY'
-          import json
-          import os
-          import sys
-          import urllib.request
-          import urllib.error
-
-          event_name = os.environ["EVENT_NAME"]
-          action_name = os.environ["ACTION_NAME"]
-          pr_author = os.environ.get("PR_AUTHOR", "")
-          issue_author = os.environ.get("ISSUE_AUTHOR", "")
-
-          # Dependabot version-update PRs are useful to merge, but not useful to broadcast.
-          if event_name == "pull_request" and pr_author == "dependabot[bot]":
-              sys.exit(0)
-
-          # Future-proofing in case other issue-producing bots get wired into this feed.
-          if event_name == "issues" and issue_author.endswith("[bot]"):
-              sys.exit(0)
-
-          payload = None
-
-          if event_name == "pull_request" and action_name == "opened":
-              payload = {
-                  "username": "MARM Get Involved",
-                  "embeds": [
-                      {
-                          "title": f"PR #{os.environ['PR_NUMBER']} opened",
-                          "url": os.environ["PR_URL"],
-                          "description": os.environ["PR_TITLE"],
-                          "color": 3447003,
-                          "fields": [
-                              {"name": "Repository", "value": os.environ["REPO"], "inline": True},
-                              {"name": "Author", "value": os.environ["ACTOR"], "inline": True},
-                          ],
-                      }
-                  ],
-              }
-
-          elif (
-              event_name == "pull_request"
-              and action_name == "closed"
-              and os.environ.get("PR_MERGED", "").lower() == "true"
-          ):
-              payload = {
-                  "username": "MARM Get Involved",
-                  "embeds": [
-                      {
-                          "title": f"PR #{os.environ['PR_NUMBER']} merged",
-                          "url": os.environ["PR_URL"],
-                          "description": os.environ["PR_TITLE"],
-                          "color": 5763719,
-                          "fields": [
-                              {"name": "Repository", "value": os.environ["REPO"], "inline": True},
-                              {"name": "Merged by", "value": os.environ["ACTOR"], "inline": True},
-                          ],
-                      }
-                  ],
-              }
-
-          elif event_name == "issues" and action_name == "opened":
-              payload = {
-                  "username": "MARM Get Involved",
-                  "embeds": [
-                      {
-                          "title": f"Issue #{os.environ['ISSUE_NUMBER']} opened",
-                          "url": os.environ["ISSUE_URL"],
-                          "description": os.environ["ISSUE_TITLE"],
-                          "color": 15844367,
-                          "fields": [
-                              {"name": "Repository", "value": os.environ["REPO"], "inline": True},
-                              {"name": "Author", "value": os.environ["ACTOR"], "inline": True},
-                          ],
-                      }
-                  ],
-              }
-
-          elif event_name == "issues" and action_name == "labeled":
-              payload = {
-                  "username": "MARM Get Involved",
-                  "embeds": [
-                      {
-                          "title": f"Issue #{os.environ['ISSUE_NUMBER']} labeled",
-                          "url": os.environ["ISSUE_URL"],
-                          "description": os.environ["ISSUE_TITLE"],
-                          "color": 10181046,
-                          "fields": [
-                              {"name": "Label", "value": os.environ["LABEL_NAME"] or "unknown", "inline": True},
-                              {"name": "Updated by", "value": os.environ["ACTOR"], "inline": True},
-                          ],
-                      }
-                  ],
-              }
-
-          if payload is None:
-              sys.exit(0)
-
-          req = urllib.request.Request(
-              os.environ["DISCORD_WEBHOOK_URL"],
-              data=json.dumps(payload).encode("utf-8"),
-              headers={"Content-Type": "application/json"},
-          )
-          try:
-              urllib.request.urlopen(req).read()
-          except urllib.error.HTTPError as exc:
-              # Discord posting is informational only; don't fail the GitHub event workflow.
-              print(f"Discord webhook post failed: HTTP {exc.code} {exc.reason}", file=sys.stderr)
-              sys.exit(0)
-          except urllib.error.URLError as exc:
-              print(f"Discord webhook post failed: {exc}", file=sys.stderr)
-              sys.exit(0)
-          PY
